Renaming concepts for clarity, added some helpers.
Changes:
Changed "actual" in the component extent wording to "scaled", to
avoid confusion where someone might think that it's where the
component actually is in screen space. "Scaled" communicates that
it's in screen space, but isn't at its final position (which is
lastRenderedExtent, the extent that includes any parent offsets
and such."

Removed "Screen" for the Component extent var names. It's unnecessary,
anything that isn't a screen-space extent can instead specify itself.

Added containsExtent() and calcClippedExtent() to Component in
preparation for upcoming clipping functionality.

Better explained the current state of scaling in ScalingHelpers.h and
made existing comments redirect to it.