Added TextBox scrolling.
Changes:
Moved Margins out to a new SharedDefinitions directory.

Added comments and a getTextOffset() to Text.

Added getters for Component's extents.

Text::calcCharacterOffset() now accounts for the textOffset.

Added TextBox::refreshTextScrollOffset(), which scrolls the Text
component to the correct offset. This lets us keep the text scrolled
to account for cursor movement, and Text's new clipping keeps it all
clipped properly.

Added a ScalingHelpers function for margins, but ended up not using
it. Might be useful later, so I'm leaving it.