Changed tick to not time itself.
Changes:
Removed timer from Screen and made it so that this library's
tick() just immediately fires the events. This puts it onto the
consumer to call the tick at the appropriate pace, but this seems
totally fine since they'll have to have some timing figured out
to render things anyway.