Started weak ref and path. Added tests.
Changes:
Moved Widget and Window to the Core directory. It makes it easier
to see what the architecture is.

Updated some comments.

Added Assert.h with an assert macro that prints.

Fixed a bug in Log.h with FenderDebugError.

Added a Catch2 submodule for unit testing.

Added a Test directory.

Added WidgetWeakRef and WidgetPath. WidgetPath is still being
worked on.

Added tests for WidgetWeakRef and WidgetPath.

Added a CMake directory with find files so a standalone unit
test build can be done.