Removed Screen constructor argument.
Changes:
Removed the Screen constructor argument from all Widgets and Windows.
It was added back when our events were subscription-based, but now
it isn't necessary for generic widgets to have a ref to the screen.
Custom project widgets and windows may still pass a ref to their
custom screen if they need to inject a dependency.