Text improvements, changed Thumbnail image order.
Changes:
Switched from the "Text" versions of the text rendering functions
to the UTF8 version, for better compatibility with fonts and
characters.

Added LCD-quality text rendering mode (requires SDL2_ttf-2.20.1).

Changed the order of the images in Thumbnail to make more sense.
The backdrop is farthest back, then the thumbnail image, then
any highlights or lowlights for states. If you want a full-size
highlight that doesn't cover the thumbnail, make the center
transparent.