AssetCache refactor, started NineSliceTexture.
Changes:
Replaced FontHandle with shared_ptr<TTF_Font>. FontHandle was only
really in place in case we needed to refactor in the future, but
the interface will likely change anyway so it isn't worth
keeping.

Moved texture asset management to this lib. Previously the consumer
would be in charge of caching textures, but doing it internally lets
us remove a lot of boilerplate AssetCache reference passing in the
consumer, and gives us better control over which textures are loaded.
Specifically, it lets us save paths instead of keeping textures in
memory when doing our multi-resolution image stuff.

Renamed loadFont() -> requestFont(), to set better expectations for
what it does.

Started adding NineSliceTexture.