Image widget refactor, added NineSliceImage.
Motivation:
To add support for more image types (nine slice specifically), a moderate
refactor had to be performed. The resulting image types provide both
simpler options for when you just want to display an image, and more
powerful options for when you need them.

Changes:
Changed Widget::renderExtent to initialize to scaledExtent. This is because
we use renderExtent for image generation, and its size needs to be non-0
in order to instantiate a texture.

Removed a lot of cases where comments or param names said "relPath", since
we changed to use full paths a while ago.

Renamed Margin -> Padding. I don't like how other engines generically use
"Margin" to mean different things (margin in a nine slice, text box margin,
etc). Padding means a specific thing, and it matches the usage of the word
in CSS. Margin also has a specific CSS-driven definition, but we don't have
any cases where we need it yet.

Added ImageType and derived classes (some are refactors of former widgets).

Refactored Image widget to be a wrapper for ImageType.