Added CollapsibleContainer, clipping fixes.
Changes:
If a widget is fully outside its given availableExtent, it now
gets its clippedExtent set to all 0 so it can be tested with
SDL_RectEmpty(). This case can occur when a widget has some
children inside of the availableExtent, and some outside of it.

Added checks for fully-clipped widgets in updateLayout() and
render() functions.

Added CollapsibleContainer.

Renamed FlowContainer -> VerticalListContainer.

Fixed Image clipping logic for when the texture extent is
smaller than the widget extent. Also added asserts.

Fixed some usage of clippedExtent in updateLayout()s,
swapping it for fullExtent. When laying out, we usually
want to use the pre-clip position instead of the clipped position.