Fixed some focus handling logic.
Changes:
Added a way to explicitly drop focus from outside of an event
handler. This is necessary because sometimes we need to drop
focus from within a user callback. An example is when closing a menu
after one of its buttons is pushed. The event isn't propagated
back to the menu after the button handles it, so the only place we
can drop focus is from within the user's button pressed callback.

Fixed an event routing bug where we were trying to find a widget to
set focus to after routing a mouse click, even if we had just dropped
focus during the handling of that click. Now, we only look for a new
focus target if the current focus target didn't change during handling.
This also fixed an inconvenience where widgets would have to handle
mouse clicks, just to maintain focus.