Added getIsTextInputFocused().
Changes:
Added a way to tell if a TextInput widget has taken keyboard focus.
This is useful since it's common in SDL2 usage to poll for held keys,
which ignores the usual event blocking mechanics.
It might be better to implement it as a "getIsFocusTaken()" or
something along those lines, to made it more broad than just TextInput
widgets. I figure it's better to wait until we reach a situation where
that kind of change is needed though, since I can't think of a need
for it at the moment.