Button improvement.
Changes:
Made it so that when a button is made invisible, it resets to its
normal state. This is for the case where pressing the button either
immediately or eventually causes the button to go invisible, while
the cursor is over it. This previously left the button in a
hovered state.