Fixed crash, incorrect Image clipping.
Changes:
Fixed a crash where, when a window had a dimension that was a
multiple of the WidgetLocator cell size (128), pointInRect() would
return true when it should have returned false. This caused an
out of bounds when getting the cell in the locator.
We seem to have been mistaken in making SDLHelpers::pointInRect().
SDL_PointInRect() correctly doesn't include x=256 when rect w=256
(rect {0, 0, 256, 256} is in pixels 0 - 255).
As such, I removed our SDLHelpers version and replaced it with
the SDL function.

Fixed incorrect clipping in the Image widget. When given a small
extent, the image was scaling to fit instead of clipping.