Started building scrollbar.
Changes:
Started building a Scrollbar helper widget to use in VerticalList-
Container, but it was tricky and not strictly necessary right now.

If we do go this direction, something to reconcile is: should the
mouse wheel logic and the scrollbar drag logic be in separate
places? If not, what's the single place that should handle both?
Other libraries handle this with a ScrollArea widget. We could
do that, or we could just build it into the scrollable containers.