Refactored updateLayout -> measure/arrange.
Changes:
Refactored updateLayout() into the more-typical measure()/arrange().
This gives us a place to handle widgets that need to change their
own logicalExtent, and lets the parents of those widgets adjust
in response.
Originally we thought we could get away with just updateLayout()
because we didn't plan on having auto-sized widgets, but with
things like CollapsibleContainer and auto-height text, we really
do need measure/arrange.