TextButton, TextInput improvements.
Changes:
Moved some measure() calls to the top of the function, since
children should be measured before the parent.

Removed a redundant alignment flag in Text.

Added auto height support to TextButton.

Fixed an issue with TextInput's offset not being calculated
correctly the first time, since it's dependent on scaledExtent
but was being called before arrange().