Added actual screen size to initialize().
Changes:
Added actual screen size to initialize() and Initializer. It's
still available as a separate call, but this lets the screen size
be accurate at widget constuction time, allowing us to avoid writing
custom arrange() functions just to get an accurate clippedExtent.
In most cases, you'll still want to write the custom arrange() so
that you properly handle window resizing, but it's nice to skip
for things like dev tools that'll never be resized during runtime.